rename sparrow_setup to faunanet_setup
