+++ tests/setup_test.py
@@ -1,4 +1,4 @@
-import faunanet.sparrow_setup as sps
+import faunanet.faunanet_setup as sps
 from pathlib import Path
 import pytest
 import shutil
